Overhauled PTMs to lookup in PSI-MOD.obo.txt
Before today, the converter for ProSight PD could only handle PTMs that I had explicitly named in the source.  Now it requires PSI-MOD.obo.txt to be present in the current working directory.  It forces to PSI-MS names whenever available and marks others with the "M:" prefix.  The code could use a more informative error when this OBO is missing, of course.